4-4 fixed the beginging bank

+++ unit4/function_test.py
@@ -9,7 +9,7 @@ def begining_bank():
 
 
 def bets ():
-    bets_v = input("Put in your bets: ")
+    bets_v = int(input("Put in your bets: "))
     #This is return the bet
     return bets_v  
 
@@ -22,9 +22,10 @@ def bets ():
 def bets_cal():
     bank_b = begining_bank()
     money = bets()
-    while (money != int) or (money > bank_b) or (money < 1):
+    while ((money != int(money)) or (money > bank_b) or (money < 1)):
         print("It has to be a whole number, not a negtive, and you need to have enogh money")
         money = bets()
+        
         
         
 bets_cal()
